+++ src/Analyzers/CSharp/DiagnosticIdentifiers.Generated.cs
@@ -211,6 +211,6 @@ namespace Roslynator.CSharp
         public const string NormalizeUsageOfInfiniteLoop = "RCS1252";
         public const string FormatDocumentationCommentSummary = "RCS1253";
         public const string NormalizeFormatOfEnumFlagValue = "RCS1254";
-        public const string RemoveEmptySyntax = "RCS1260";
+        public const string RemoveEmptySyntax = "RCS1255";
     }
 }