+++ src/Tests/Analyzers.Tests/RCS1255RemoveEmptySyntaxTests.cs
@@ -8,7 +8,7 @@ using Xunit;
 
 namespace Roslynator.CSharp.Analysis.Tests
 {
-    public class RCS1260RemoveEmptySyntaxTests : AbstractCSharpDiagnosticVerifier<RemoveEmptySyntaxAnalyzer, RemoveEmptySyntaxCodeFixProvider>
+    public class RCS1255RemoveEmptySyntaxTests : AbstractCSharpDiagnosticVerifier<RemoveEmptySyntaxAnalyzer, RemoveEmptySyntaxCodeFixProvider>
     {
         public override DiagnosticDescriptor Descriptor { get; } = DiagnosticRules.RemoveEmptySyntax;